refactor: export mods menu component to dedicated file
